requirements.txt
pandas==2.1.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openpyxl==3.1.2

# API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

